Added flask1.py The tutorial forgot to mention adding the #! at the top of this python file.
It gives some security warnings related to the secret key and the debug flag.
Apparently, if you leave the debug mode activated in a "production server"
(not sure what that is yet), it will let users execute code on the server.
They also warn to keep the secret key secret.